refactor layout page and use org context provider
