+++ app/[[...slug]]/page.tsx
@@ -1,8 +1,5 @@
-import { preloadQuery } from "convex/nextjs";
-import { api } from "../../convex/_generated/api";
-import { headers } from "next/headers";
-import Home from "../components/Home";
 import { Id } from "@/convex/_generated/dataModel";
+import Feed from "../components/Feed";
 
 export default async function App({
   params,
@@ -17,14 +14,5 @@ export default async function App({
     feedId = slug[1] as Id<"feeds">;
   }
 
-  const orgSubdomain = (await headers()).get("x-org-host") || "";
-
-  const preloadedOrg = await preloadQuery(
-    api.organizations.getOrganizationBySubdomain,
-    {
-      subdomain: orgSubdomain,
-    }
-  );
-
-  return <Home preloadedOrg={preloadedOrg} feedId={feedId} />;
+  return <Feed feedIdSlug={feedId} />;
 }
