+++ app/layout.tsx
@@ -2,7 +2,12 @@ import type { Metadata } from "next";
 import { Lato, Gentium_Plus } from "next/font/google";
 import "./globals.css";
 import { ClerkProvider } from "@clerk/nextjs";
-import ConvexClientProvider from "./ConvexClientProvider";
+import ConvexClientProvider from "./context/ConvexClientProvider";
+import { OrganizationProvider } from "./context/OrganizationProvider";
+import { api } from "../convex/_generated/api";
+import { preloadQuery } from "convex/nextjs";
+import { headers } from "next/headers";
+import OrganizationLayout from "./components/OrganizationLayout";
 
 const lato = Lato({
   variable: "--font-lato",
@@ -22,16 +27,29 @@ export const metadata: Metadata = {
   description: "Your feed",
 };
 
-export default function RootLayout({
+export default async function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const orgSubdomain = (await headers()).get("x-org-host") || "";
+
+  const preloadedOrg = await preloadQuery(
+    api.organizations.getOrganizationBySubdomain,
+    {
+      subdomain: orgSubdomain,
+    }
+  );
+
   return (
     <html lang="en">
       <body className={`${lato.variable} ${gentiumPlus.variable}`}>
         <ClerkProvider>
-          <ConvexClientProvider>{children}</ConvexClientProvider>
+          <ConvexClientProvider>
+            <OrganizationProvider organization={preloadedOrg}>
+              <OrganizationLayout>{children}</OrganizationLayout>
+            </OrganizationProvider>
+          </ConvexClientProvider>
         </ClerkProvider>
       </body>
     </html>
